Website Beta v1.1
Adjusted sizing of the images on the Portfolio Page to look better.
Updated the title of the index page.
Added a new art img and removed an old one.
Updated info with updated data in her google drive.
Adjusted formatting for multiple sections of the webpages.
Fixed minor issues in the code.